feat: Extract CrossAppLoginViewModelFactory to its own file
